+++ opencascade-hs/cpp/hs_Geom_Curve.h
@@ -9,6 +9,8 @@ extern "C" {
 void hs_delete_Handle_Geom_Curve(Handle(Geom_Curve) * handle);
 
 gp_Pnt * hs_Geom_Curve_value(Handle(Geom_Curve) * curve, double u);
+
+gp_Vec * hs_Geom_Curve_dn(Handle (Geom_Curve) * curve, double u, int n);
 #ifdef __cplusplus
 }
 #endif